+++ custom_components/powercalc/const.py
@@ -26,6 +26,7 @@ CONF_MULTIPLY_FACTOR = "multiply_factor"
 CONF_MULTIPLY_FACTOR_STANDBY = "multiply_factor_standby"
 CONF_MIN_WATT = "min_watt"
 CONF_MAX_WATT = "max_watt"
+CONF_POWER_FACTOR = "power_factor"
 CONF_POWER_SENSOR_NAMING = "power_sensor_naming"
 CONF_POWER = "power"
 CONF_MIN_POWER = "min_power"
